Refactored search folder, adding logging config
