+++ firstvoices/backend/admin/dictionary_admin.py
@@ -120,3 +120,4 @@ admin.site.register(DictionaryEntryRelatedCharacter, HiddenBaseAdmin)
 admin.site.register(Translation, HiddenBaseAdmin)
 admin.site.register(AlternateSpelling, HiddenBaseAdmin)
 admin.site.register(Pronunciation, HiddenBaseAdmin)
+admin.site.register(WordOfTheDay, HiddenBaseAdmin)
