+++ firstvoices/backend/search/indices/dictionary_entry_document.py
