+++ firstvoices/backend/urls.py
@@ -17,10 +17,10 @@ from backend.views.word_of_the_day_views import WordOfTheDayView
 # app-level APIs
 ROUTER = DefaultRouter(trailing_slash=True)
 ROUTER.register(r"user", UserViewSet, basename=r"user")
+ROUTER.register(r"search", CustomSearchViewSet, basename="search")
 ROUTER.register(r"parts-of-speech", PartsOfSpeechViewSet, basename="partofspeech")
 ROUTER.register(r"my-sites", MySitesViewSet, basename="my-sites")
 ROUTER.register(r"sites", SiteViewSet, basename="site")
-ROUTER.register(r"search", CustomSearchViewSet, basename="search")
 
 # site-level APIs
 sites_router = NestedSimpleRouter(ROUTER, r"sites", lookup="site")
